requirements.txt
# -f https://download.pytorch.org/whl/cpu/torch_stable.html
pathlib==1.0.1
streamlit>=1.5
streamlit-drawable-canvas>=0.8
torch==0.10.2+cpu
torchvision==0.11.3+cpu
matplotlib==3.5.1
matplotlib-inline==0.1.3
numpy==1.22.2
Pillow==9.0.1

gdown==4.3.1
